Pushes API and pull-to-refresh

+++ Plutocrat/Model/UserManager.h
@@ -14,7 +14,7 @@
 + (void)removeUser;
 + (BOOL)isLogin;
 + (BOOL)isDefeated;
-+ (int)currentUserId;
++ (NSUInteger)currentUserId;
 + (NSString *)getHeader:(NSString *)key;
 + (BOOL)hasLastLogin;
 + (int)lastLogin;
